2013/brp 2013-02-07/BRP/branches/0.2.1-dual-jdbc/business/src/main/java/nl/bzk/brp/business/bedrijfsregels/impl/migratie/BRBY0012.java
/**
 * This file is copyright 2017 State of the Netherlands (Ministry of Interior Affairs and Kingdom Relations).
 * It is made available under the terms of the GNU Affero General Public License, version 3 as published by the Free Software Foundation.
 * The project of which this file is part, may be found at https://github.com/MinBZK/operatieBRP.
 */

package nl.bzk.brp.business.bedrijfsregels.impl.migratie;

import java.util.ArrayList;
import java.util.List;

import nl.bzk.brp.business.bedrijfsregels.ActieBedrijfsRegel;
import nl.bzk.brp.model.basis.Identificeerbaar;
import nl.bzk.brp.model.objecttype.logisch.Actie;
import nl.bzk.brp.model.objecttype.logisch.Persoon;
import nl.bzk.brp.model.objecttype.operationeel.statisch.Soortmelding;
import nl.bzk.brp.model.validatie.Melding;
import nl.bzk.brp.model.validatie.MeldingCode;
import nl.bzk.brp.util.DatumUtil;


/**
 * Datum einde geldigheid mag niet groter zijn dan actuele datum.
 *
 * @brp.bedrijfsregel BRBY0012
 *
 */
public class BRBY0012 implements ActieBedrijfsRegel<Persoon> {

    @Override
    public String getCode() {
        return "BRBY0012";
    }

    @Override
    public List<Melding> executeer(final Persoon huidigeSituatie, final Persoon nieuweSituatie,
            final Actie actie)
    {
        List<Melding> meldingen = new ArrayList<Melding>();

        if (actie.getDatumEindeGeldigheid() != null && DatumUtil.vandaag().voor(actie.getDatumEindeGeldigheid())) {
            meldingen.add(new Melding(Soortmelding.FOUT, MeldingCode.BRBY0012,
                    (Identificeerbaar) actie, "datumEindeGeldigheid"));
        }

        return meldingen;
    }
}
